bloom2/bloom_python/requirements.txt
# requirements.txt
firebase-admin>=6.0.1
pyserial>=3.5